test: update tests
